feat: complete u64 migration + smart_http error enum
u64 migration (all remaining fields):
- FileChunkRecord.range_start/range_end: u32→u64 (index/record.rs)
- ReconstructionChunkRange.start/end: u32→u64 (xet_adapter/model.rs)
- XorbObjectInfoV1: num_chunks, offsets, lengths all u32→u64
- XorbObject.info_length: u32→u64
- XorbObject format V1→V2 with backward compat in deserialize
- All callers updated across 19 files

smart_http error enum:
- Created SmartHttpError with 34 thiserror variants
- Replaced 11 map_err(format!(...)) + 26 other string errors
- Updated 5 function signatures from Result<_, String> to Result<_, SmartHttpError>
- 520 hub_api tests pass

7,963 tests pass, CI green

+++ crates/fsck/src/types.rs
@@ -111,9 +111,9 @@ pub enum FsckIssueDetail {
     #[error("xorb range {range_start}..{range_end} exceeded {chunk_count} chunks")]
     XorbRangeExceededChunkCount {
         /// Requested range start.
-        range_start: u32,
+        range_start: u64,
         /// Requested range end.
-        range_end: u32,
+        range_end: u64,
         /// Available chunk count.
         chunk_count: usize,
     },
